Halo Reach Tag File Writer WIP 2

+++ Tags/BlamboozleLib/halo4/h4_tag_reference.cpp
@@ -2,7 +2,7 @@
 
 c_h4_tag_reference::c_h4_tag_reference(const char* h4_data, const s_h4_tag_reference_definition* reference_definition, unsigned long definition_address) :
 	reference_definition(reference_definition),
-	flags(reference_definition->flags),
+	flags(static_cast<unsigned long>(reference_definition->flags)),
 	public_symbol()
 {
 	ASSERT(BCS_SUCCEEDED(symbol_manager_get_public_symbol_by_rva_plus_base(h4_data, definition_address, public_symbol)));
